chore: refactor kustomize path related funcs
